Some folder refactoring.
