+++ client/src/components/service_offers/Candidate.js
